aurorawatch-pipeline/requirements.txt
pylint
pytest
requests
pytest-cov
psycopg2-binary
python-dotenv